enumUpdate

+++ src/main/java/bike/service/app/service/userroles/MechanicService.java
@@ -10,7 +10,6 @@ import bike.service.app.model.Bike;
 import bike.service.app.model.Client;
 import bike.service.app.model.Order;
 import bike.service.app.model.Services;
-import bike.service.app.model.Users;
 import bike.service.app.model.repository.BikeRepository;
 import bike.service.app.model.repository.ClientRepository;
 import bike.service.app.model.repository.OrderRepository;
@@ -32,14 +31,14 @@ public class MechanicService {
     private UsersService userService;
 
     public void editOrderById(String edit, int id) {
-       Optional<Order> orderOptional = orderRepository.findById(id);
-       if(orderOptional.isPresent()) {
-        Order newOrder = orderOptional.get();
-        newOrder.setService(edit);
+        Optional<Order> orderOptional = orderRepository.findById(id);
+        if (orderOptional.isPresent()) {
+            Order newOrder = orderOptional.get();
+            newOrder.setService(edit);
 
-        orderRepository.save(newOrder);
+            orderRepository.save(newOrder);
 
-       }
+        }
     }
 
     public void deleteOrderById(int id) {
@@ -48,42 +47,42 @@ public class MechanicService {
         List<Bike> bikes = bikeRepository.findAll();
         List<Services> services = servicesRepository.findAll();
 
-        for (Client client : clients) {
-            clientRepository.deleteById(client.getClientId());
-            client.getOrder();
-        }
+        if (order.getStatus().equals(Order.Status.DONE)) {
+            for (Client client : clients) {
+                clientRepository.deleteById(client.getClientId());
+                client.getOrder();
+            }
 
-        for (Bike bike : bikes) {
-            bike.getOrder();
-            bikeRepository.deleteById(bike.getBikeId());
-        }
+            for (Bike bike : bikes) {
+                bike.getOrder();
+                bikeRepository.deleteById(bike.getBikeId());
+            }
 
-        for (Services service : services) {
-            service.getOrder();
-            clientRepository.deleteById(service.getServiceId());
+            for (Services service : services) {
+                service.getOrder();
+                clientRepository.deleteById(service.getServiceId());
 
-        }
+            }
 
-        if (!(order.getOrderId() == 0)) {
-            orderRepository.deleteById(id);
-            System.out.println("order deleted");
+            if (!(order.getOrderId() == 0)) {
+                orderRepository.deleteById(id);
+                System.out.println("order deleted");
+            }
+        } else {
+            throw new RuntimeException("Order status ACTIVE");
         }
     }
 
-    public Order saveMechanicToOrder(Order order, int id) {
-        System.out.println("saveMechanicToOrder");
-
-        Users mechanic = userService.getUserById(id);
-
-        order.setMechanic(mechanic);
-        System.out.println("Order before save: " + order);
-
-        orderRepository.save(order);
-        System.out.println("Order after save: " + order);
-        // mechanicRepository.save(mechanic);
-        return order;
-    }
-
+    // public Order saveMechanicToOrder(Order order, int id) {
+    //     System.out.println("saveMechanicToOrder");
+    //     Users mechanic = userService.getUserById(id);
+    //     order.setMechanic(mechanic);
+    //     System.out.println("Order before save: " + order);
+    //     orderRepository.save(order);
+    //     System.out.println("Order after save: " + order);
+    //     // mechanicRepository.save(mechanic);
+    //     return order;
+    // }
     public Order getOrderById(int id) {
         Optional<Order> order = orderRepository.findById(id);
         return order.get();
